pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argus-panoptes"
version = "0.1.0"
description = "The all-seeing eye for your AI agents — unified observability across any agent framework."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "AJ Geddes" },
]
keywords = ["ai", "agents", "observability", "metrics", "opentelemetry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "sqlmodel>=0.0.22",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "watchdog>=6.0",
    "httpx>=0.28",
    "pydantic>=2.10",
    "sse-starlette>=2.2",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "bandit>=1.8",
    "httpx",
    "pre-commit>=4.0",
]

[project.urls]
Homepage = "https://github.com/aj-geddes/argus-panoptes"
Repository = "https://github.com/aj-geddes/argus-panoptes"

[tool.hatch.build.targets.wheel]
packages = ["src/argus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning:sqlalchemy.*",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"src/argus/schemas/otlp.py" = ["N815"]  # camelCase field names match OTLP spec

[tool.ruff.lint.isort]
known-first-party = ["argus"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["sqlmodel.*", "watchdog.*", "sse_starlette.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/argus"]
omit = ["*/migrations/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

[tool.bandit]
exclude_dirs = ["tests"]
